+++ spec/todoist_spec.rb
@@ -129,6 +129,14 @@ PATCH
         expect(failures).to be_empty
         expect(markdowns).to be_empty
       end
+
+      it "does not raise when git returns nil" do
+        invalid = [nil]
+        allow(@dangerfile.git).to receive(:modified_files).and_return(invalid)
+        allow(@dangerfile.git).to receive(:added_files).and_return([])
+
+        expect { @todoist.warn_for_todos }.to_not raise_error
+      end
     end
   end
 end